role auth wip

+++ Group5/routes/web.php
@@ -52,27 +52,27 @@ Route::get('/register-view', function(){
     return view('auth.register');
 });
 
-//users CRUD
-Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update');
-Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show');
-Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy');
+//Admin CRUD
+Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit')->middleware('role:admin');
+Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update')->middleware('role:admin');
+Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show')->middleware('role:admin');
+Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy')->middleware('role:admin');
 
 
 //Employer
-Route::get('/jobs', [JobController::class, 'index'])->name('job.index');
-Route::get('/jobs/create', [JobController::class, 'create'])->name('job.create');
-Route::post('/store', [JobController::class, 'store'])->name('job.store');
+Route::get('/jobs', [JobController::class, 'index'])->name('job.index')->middleware('role:employer');
+Route::get('/jobs/create', [JobController::class, 'create'])->name('job.create')->middleware('role:employer');
+Route::post('/store', [JobController::class, 'store'])->name('job.store')->middleware('role:employer');
 Route::get('/job/{job}/edit', [JobController::class, 'edit'])->name('job.edit');
-Route::put('/job/{job}/update', [JobController::class, 'update'])->name('job.update');
-Route::delete('/job/{job}/delete', [JobController::class, 'delete'])->name('job.delete');
+Route::put('/job/{job}/update', [JobController::class, 'update'])->name('job.update')->middleware('role:employer');
+Route::delete('/job/{job}/delete', [JobController::class, 'delete'])->name('job.delete')->middleware('role:employer');
 
 //applicant
-Route::get('/applicant/jobs', [JobController::class, 'showAllJobsToApplicant'])->name('applicant.index');
-Route::get('/applicant/{job}/apply', [ApplicationController::class, 'apply'])->name('applicant.apply');
+Route::get('/applicant/jobs', [JobController::class, 'showAllJobsToApplicant'])->name('applicant.index')->middleware('role:applicant');
+Route::get('/applicant/{job}/apply', [ApplicationController::class, 'apply'])->name('applicant.apply')->middleware('role:applicant');
 
-Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat');
-Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store');
-Route::get('/category/edit/{id}', [CategoryController::class, 'Edit']);
-Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category');
-Route::get('/category/delete/{id}', [CategoryController::class, 'Delete']);
+Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat')->middleware('role:admin');
+Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store')->middleware('role:admin');
+Route::get('/category/edit/{id}', [CategoryController::class, 'Edit'])->middleware('role:admin');
+Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category')->middleware('role:admin');
+Route::get('/category/delete/{id}', [CategoryController::class, 'Delete'])->middleware('role:admin');
